+++ src-docs/src/views/markdown_editor/markdown_editor_with_plugins.js
@@ -293,6 +293,7 @@ export default () => {
     <>
       <EuiMarkdownEditor
         aria-label="EUI markdown editor with plugins demo"
+        placeholder="Your markdown here..."
         value={value}
         onChange={setValue}
         height={400}